requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
certifi==2023.7.22
pydantic==2.5.0